refactor(logging): move logging setup into src/logging/ module folder
Consolidates the structured-JSON logging backend (added in #814) and its
tests into a dedicated module folder, matching the src/utils and
src/paths convention (mod.rs + tests.rs, wired via #[path]).

Closes #852
