+++ src/logging/tests.rs
